requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyYAML==5.1.2
soupsieve==1.9.3
Werkzeug==0.15.5
